Another fix for sorting by quantity
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@414 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/pcgen/gui/tabs/InfoGear.java
@@ -3424,7 +3424,7 @@ public final class InfoGear extends FilterAdapterPanel implements CharacterInfoT
 				case COL_QTY:
 					if (eq != null)
 					{
-						retVal = new Double(eq.qty());
+						retVal = new Float(eq.qty());
 					}
 					break;
